backend/pyproject.toml
[project]
name = "todo-ai-chatbot-backend"
version = "0.1.0"
description = "AI-powered todo management backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.100.0",
    "sqlmodel>=0.0.8",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "mcp>=0.2.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]